melhorias exportação de notificações
