+++ frontend/src/services/NotificacaoService.js
@@ -32,8 +32,16 @@ export default {
     dataInicial, dataFinal, dataEvolucaoInicial, dataEvolucaoFinal,
   }) {
     const time = new Date().getTime();
-    return DownloadService.downloadFile(`/notificacao-gera-excel?dataInicial=${dataInicial}&dataFinal=${dataFinal}
+    return DownloadService.downloadFile(`/exportar/csv?dataInicial=${dataInicial}&dataFinal=${dataFinal}
     &dataEvolucaoInicial=${dataEvolucaoInicial}&dataEvolucaoFinal=${dataEvolucaoFinal}`,
     `notificacoes-${time}.xlsx`);
+  },
+  downloadNotificacoesCSV({
+    dataInicial, dataFinal, dataEvolucaoInicial, dataEvolucaoFinal,
+  }) {
+    const time = new Date().getTime();
+    return DownloadService.downloadFile(`/exportar/csv?dataInicial=${dataInicial}&dataFinal=${dataFinal}
+    &dataEvolucaoInicial=${dataEvolucaoInicial}&dataEvolucaoFinal=${dataEvolucaoFinal}`,
+    `notificacoes-${time}.csv`);
   },
 };
